Use glTexStorage2D
From the Adreno development guide:

[mutable textures] greatly reduced the optimization possibilities for driver implementations, which were forced to keep
track of the completeness of all textures used by the application. Texture completeness
verification is a significant overhead, especially if it has to be executed for every draw call the
application makes.

On Adreno platforms, the use of immutable textures has a major performance advantage. Always
use immutable textures and avoid all mutable textures.