Don't invalidate framebuffer attachments.
This was put in place to be a bit of a performance hack
It also happened to fix some driver bugs on certain devices, but it was causing some graphical glitches on Android
For instance, the portrait in Super Smash that shows up when you go off screen had no image in it.

This should probably still be implemented in GLideN64, but it needs to be implemented in the right circumstances.