Reduce calls to glBindFramebuffer
GLideN64 makes redundant calls to glDrawBuffers and glFramebufferTexture2D.

Both of those things are part of the state of the framebuffer, so they only need to be called once per framebuffer.
GLideN64 calls them at the start of each frame, those calls themselves aren't terrible, but they require a call to glBindFramebuffer before they are called, which is redundant and bad for performance on mobile.

Also, glBlitFramebuffer requires calls to glBindFramebuffer before it is used, but there is an alternative (if the device supports it), glCopyImageSubData.
This function can copy an FBO texture without having to call glBindFramebuffer if the source and destination sizes are the same.

In games that make heavy use of FBOs these changes can result in a dramatic reduction of calls to glBindFramebuffer.

Mario Tennis and Conkers: ~20 glBindFramebuffer calls per frame becomes ~10