Controller changes, please read.
Right now RetroPad A=N64 A and RetroPad B=N64 B, no more.

Now, RetroPad Y=B, and RetroPad B=A.

This is more intuitive for most people, since this is how the original N64 controller was laid out.

The only downside is that folks with replica N64 controllers will need to remap them, since it only has an "A" and "B" button and they are mapped to RetroPad A&B

Most people use 4 button controllers, so I think this is a good change.