Clean up properly on deinit to allow reinitialization
The two global variables we care about cleaning up properly, that are
not currently, are l_CoreInit (only reset by CoreShutdown, which was
not previously called from anywhere) and l_ROMOpen. l_ROMOpen is
interesting because it is reset by M64CMD_ROM_CLOSE which is called
from retro_unload_game; but sadly this is too early for it to be
called, as g_EmulatorRunning is still 1. So instead we have to wait
for the emulator to have been stopped by M64CMD_STOP. Further, if the
current_rdp_type is not RDP_PLUGIN_GLIDEN64, M64CMD_ROM_CLOSE is not
invoked until retro_deinit.